Addressed comments for DELETE Lifecycle API

+++ api/pkg/s3/bucketlifecycledelete.go
@@ -15,36 +15,51 @@
 package s3
 
 import (
+	"net/http"
+
 	"github.com/emicklei/go-restful"
 	"github.com/micro/go-log"
 	"github.com/opensds/multi-cloud/api/pkg/policy"
 	s3 "github.com/opensds/multi-cloud/s3/proto"
 	"golang.org/x/net/context"
-	"net/http"
 )
 
 func (s *APIService) BucketLifecycleDelete(request *restful.Request, response *restful.Response) {
 	if !policy.Authorize(request, response, "bucket:delete") {
 		return
 	}
+	var foundID int
+	resp := make([]s3.BaseResponse, 0)
 	bucketName := request.PathParameter("bucketName")
 	ruleID := request.Request.URL.Query()["ruleID"]
-
-	ctx := context.Background()
-	log.Logf("Received request for bucket lifecycle delete for bucket: %s and the ruleID: %s", bucketName, ruleID)
-	bucket, _ := s.s3Client.GetBucket(ctx, &s3.Bucket{Name: bucketName})
-	for _, lcRule := range bucket.LifecycleConfiguration {
-		for _, id := range ruleID {
-			if lcRule.ID == id {
-				deleteInput := s3.DeleteLifecycleInput{Bucket: bucketName, RuleID: id}
-				res1, err := s.s3Client.DeleteBucketLifecycle(ctx, &deleteInput)
-				if err != nil {
-					response.WriteError(http.StatusInternalServerError, err)
-					return
+	if ruleID != nil {
+		ctx := context.Background()
+		log.Logf("Received request for bucket lifecycle delete for bucket: %s and the ruleID: %s", bucketName, ruleID)
+		bucket, _ := s.s3Client.GetBucket(ctx, &s3.Bucket{Name: bucketName})
+		for _, lcRule := range bucket.LifecycleConfiguration {
+			for _, id := range ruleID {
+				if lcRule.ID == id {
+					foundID = foundID+1
+					deleteInput := s3.DeleteLifecycleInput{Bucket: bucketName, RuleID: id}
+					res1, err := s.s3Client.DeleteBucketLifecycle(ctx, &deleteInput)
+					if err != nil {
+						response.WriteError(http.StatusInternalServerError, err)
+						return
+					}
+					resp = append(resp, *res1)
 				}
-				response.WriteEntity(res1)
 			}
 		}
+		if foundID == 0 || (foundID < len(ruleID)){
+			response.WriteErrorString(http.StatusBadRequest, WrongRuleIDForLifecycleDelete)
+			return
+		}
+		if resp != nil {
+			response.WriteEntity(resp[0])
+		}
+	} else {
+		response.WriteErrorString(http.StatusBadRequest, NoRuleIDForLifecycleDelete)
+		return
 	}
 	log.Log("Delete bucket lifecycle successfully.")
 }
